+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Build project to esm (e.s. modules) and cjs (commonjs)
 - Codecov
+- Array module
 
 ### Changed
 - Update README